third_party/blink/web_tests/virtual/sanitizer-api/external/wpt/sanitizer-api/sanitizer-names.https-expected.txt

This is a testharness.js-based test.
[FAIL] Element names in config item: elements
  sanitizer.getConfiguration is not a function
[FAIL] Element names in config item: removeElements
  sanitizer.getConfiguration is not a function
[FAIL] Element names in config item: replaceWithChildrenElements
  sanitizer.getConfiguration is not a function
[FAIL] Attribute names in config item: attributes
  sanitizer.getConfiguration is not a function
[FAIL] Attribute names in config item: removeAttributes
  sanitizer.getConfiguration is not a function
[FAIL] Namespaced elements #1: elements: ["svg"]
  assert_equals: expected "Hello" but got ""
[FAIL] Namespaced elements #3: elements: ["math"]
  assert_equals: expected "Hello" but got ""
[FAIL] Namespaced elements #4: elements: [{"name":"math","namespace":"http://www.w3.org/2000/svg"}]
  assert_equals: expected "Hello" but got ""
[FAIL] Lower-case element names #0: "svg:feblend"
  assert_equals: expected "<feBlend></feBlend>" but got ""
[FAIL] Mixed case element names #0: "feBlend" is preserved in config.
  (intermediate value).getConfiguration is not a function
[FAIL] Lower-case element names #1: "svg:fecolormatrix"
  assert_equals: expected "<feColorMatrix></feColorMatrix>" but got ""
[FAIL] Mixed case element names #1: "feColorMatrix" is preserved in config.
  (intermediate value).getConfiguration is not a function
[FAIL] Lower-case element names #2: "svg:textpath"
  assert_equals: expected "<textPath></textPath>" but got ""
[FAIL] Mixed case element names #2: "textPath" is preserved in config.
  (intermediate value).getConfiguration is not a function
Harness: the test ran to completion.
